Fix systems count per issue on Compliance
If you have a system with multiple policies and the failed rules are the
same in multiple policies, the systems count in the Remediations column
will be counted wrong, because the same system is listed multiple times.

This shouldn't change any functionality, just the step of the
remediations wizard where it shows the systems count.