RHICOMPL-572 - Prevent rule overflow in Compliance Rules table
The Stack Patternfly component is display: flex and therefore it's
height is set as 0. This makes it impossible for the Table component to
understand how tall is each of the rows.

Instead, I'm switching the rows to use a simple TextContent component
which the Table can effortlessly use to figure out the correct height.