Fix notifications babel config for direct imports
Many parts of the documentation suggests importing the `addNotification` action directly.

However, that didn't work with the babel-plugin-transform-imports configuration offered in the README file.
This commit fixes the documented configuration to allow direct import of `addNotification` and `removeNotification`, and adds a named export `notifications` to the reducer file to make this suggested configuration work.