fix(ComplianceRemediationButton): RHICOMPL-873 color
Fixes color of ComplianceRemediationButton to match button primary color
(white).  The color is also used for on disabled state.

In general scenario the icon within a button would share the same color
as the text.  Unfortunately, this is not the case (?).

Use of direct CSS variables is deliberate for consistency.  Previously
used background color (provided from react-tokens) had a wrong color.