Replace PF navigation with styled default HTML components
There is an issue with the PF nav. It renders itslef about 140 times if
any action is taken on the page (for example expanding the code
preview). Not sure what the issue is just yet, but for now we will use
native HTML elements with PF styling to avoid the excesive rendering
problem.