Fix webpack dev server container loading (#1314)
* Remove webpack optimization config

The optimization configuration entry was causing issues in local
development environment. It was unable to initialize the self-referencing remote container. The specific configuration that was causing the issue is the [`splitchunks all`](https://webpack.js.org/plugins/split-chunks-plugin/#splitchunkschunks) setting. We can remove the split chunks config as module federations add additional splitting behavior.

* Do not inject JS via HTML webpack plugin.

The HTML template will not be used and we should top using it to prevent
template difference between development mode and production env.