feat: Add Open Source Badge component
This commit adds a new component, OpenSourceBadge.

This badge is already in use by Image Builder
(https://console.redhat.com/insights/image-builder). It was created with
two purposes in mind:
1.) To ensure that consoledot users know that the application they are
  using is open source.
2.) To allow developers (both internal and external) to quickly find
  relevant repositories, whether they wish to contribute or simply need
  to inspect the code.

Adding this component to frontend-components will allow it to be re-used
by other applications on consoledot, ensuring a consistent user
experience across the platform.