feat: add single select filter component
This PR adds a single select filter dropdown component for use with the
Conditional Filter. It is used on conjuction with the `singleSelect`
conditional filter type. Docs are included in the PR.

To convert a radio filter to single select, you should be able to simply
change the type from `radio` to `singleSelect`.