feat(config): add new dev-proxy cmd to fec bin
This adds a new "dev-proxy" command to the FEC binary in the config
package. This new command uses the new and better
frontend-development-proxy which is based on Caddy, where the main
benefits alongside speed are HTTP2 support and better websocket support.
The new command imitates the behavior of the dev command as much as
possible so it should be a drop-in replacement for the consumers.